+++ packages/terra-date-time-picker/src/DateTimePicker.jsx
@@ -64,6 +64,11 @@ const propTypes = {
    */
   onChangeRaw: PropTypes.func,
   /**
+   * A callback function to execute when a selection is made in the date picker.
+   * The first parameter is the event. The second parameter is the selected input value in ISO format.
+   */
+  onSelect: PropTypes.func,
+  /**
    * A callback function to let the containing component (e.g. modal) to regain focus.
    */
   releaseFocus: PropTypes.func,
@@ -97,6 +102,7 @@ const defaultProps = {
   minDateTime: undefined,
   onChange: undefined,
   onChangeRaw: undefined,
+  onSelect: undefined,
   releaseFocus: undefined,
   requestFocus: undefined,
   timeInputAttributes: undefined,
@@ -163,6 +169,9 @@ class DateTimePicker extends React.Component {
 
     if (!previousDateTime || previousDateTime.format() !== updatedDateTime.format()) {
       this.checkAmbiguousTime(updatedDateTime);
+    }
+    if (this.props.onSelect) {
+      this.props.onSelect(event, updatedDateTime.format());
     }
   }
 
@@ -402,6 +411,7 @@ class DateTimePicker extends React.Component {
       includeDates,
       onChange,
       onChangeRaw,
+      onSelect,
       maxDateTime,
       minDateTime,
       name,