+++ packages/terra-time-input/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.7.0 - (April 24, 2019)
+------------------
+### Changed
+* Moved packages we use solely in doc examples and tests to devDependencies
+
 3.6.0 - (March 26, 2019)
 ------------------
 ### Changed